dvc-wine/pyproject.toml
[project]
name = "dvc-wine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc[s3]>=3.59.2",
    "joblib>=1.5.0",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
]
